Add files via upload
This project was an incredible journey of exploring the power of Machine Learning and Python libraries. I used numpy and pandas for data handling and manipulation, and scikit-learn for model building and evaluation.

The core of the project involved feature extraction, accuracy scoring, and logistic regression from the linear model. I also used TfidfVectorizer for text feature extraction and cosine_similarity for determining similarity between different movies.

The result? A system that can recommend movies based on user preferences with impressive accuracy! 🌟

I’m grateful for the learning experience and look forward to applying these skills in future projects. If you’re interested in learning more or collaborating on similar projects, feel free to connect!